Set correct fetch size to enable streaming in MySQL result sets.
The MySQL Connector/J only streams result sets if statements are
configured with a fetch size of {@code Integer.MIN_VALUE} and the
options TYPE_FORWARD_ONLY and CONCUR_READ_ONLY. See
https://dev.mysql.com/doc/connector-j/en/connector-j-reference-implementation-notes.html
for details. Since this fetch size setting does not seem to be specified
JDBC behaviour, we handle it as a special case.