Makes the record ids emitted by SqlStatementPipe configurable
In commit 93d6fd19391039595033c14a936b72ebd69940f7 a change was introduced to
use the values of one column in the result set as record ids. The same commit
added a setter to SqlStreamSource to allow the user to control which column
is used as the source of the record ids. This commit adds the same functionality
to SqlStatementPipe.