Refactored constructors in the *Query classes
Removed the JNDI datasource support from BaseQuery. Users of the *Query classes
now have to aquire a database connection themselves. JdbcUtil provides a utility
function (getConnection) to do this.

This commit reduces the number of constructors in the *Query classes and also
helps to simplify statement/query instantiation in the Sql* modules.