Fixes #5: MySQL fails to return result set for `getGeneratedKeys`
The MySQL connector seems to use the same result set in `getResultSet`
and `getGeneratedKeys`. Once the result set received from the former
method is closed, a call to the latter method fails because the result
set is already closed.

This commit works around this issue by keeping the original result set
open until after emitting the generated keys.